move `createUITemplate()` override to FuelMultiblockController
