+++ src/main/java/gregtech/api/metatileentity/multiblock/FuelMultiblockController.java
@@ -4,10 +4,15 @@ import gregtech.api.GTValues;
 import gregtech.api.capability.IEnergyContainer;
 import gregtech.api.capability.impl.EnergyContainerList;
 import gregtech.api.capability.impl.MultiblockFuelRecipeLogic;
+import gregtech.api.gui.GuiTextures;
+import gregtech.api.gui.ModularUI;
+import gregtech.api.gui.widgets.AdvancedTextWidget;
+import gregtech.api.gui.widgets.ImageCycleButtonWidget;
 import gregtech.api.recipes.RecipeMap;
 import gregtech.api.util.GTUtility;
 import gregtech.api.util.TextFormattingUtil;
 import gregtech.common.ConfigHolder;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.Style;
@@ -72,6 +77,22 @@ public abstract class FuelMultiblockController extends RecipeMapMultiblockContro
                 textList.add(new TextComponentTranslation("gregtech.multiblock.idling"));
             }
         }
+    }
+    @Override
+    protected ModularUI.Builder createUITemplate(EntityPlayer entityPlayer) {
+        ModularUI.Builder builder = new ModularUI.Builder(GuiTextures.BACKGROUND, 176 + 90, 216);
+        builder.image(7, 4, 162 + 90, 121 + 6, GuiTextures.DISPLAY);
+        builder.label(13, 9, getMetaFullName(), 0xFFFFFF);
+        builder.widget(new AdvancedTextWidget(13, 19, this::addDisplayText, 0xFFFFFF)
+                .setMaxWidthLimit(156 + 90)
+                .setClickHandler(this::handleDisplayClick));
+        if (shouldShowVoidingModeButton()) {
+            builder.widget(new ImageCycleButtonWidget(149, 121 - 17, 18, 18, GuiTextures.BUTTON_VOID_MULTIBLOCK,
+                    4, this::getVoidingMode, this::setVoidingMode)
+                    .setTooltipHoverString(MultiblockWithDisplayBase::getVoidingModeTooltip));
+        }
+        builder.bindPlayerInventory(entityPlayer.inventory, GuiTextures.SLOT, 7 + 45, 134);
+        return builder;
     }
 
     @Nonnull
